refactor: tweak net device via xpath
xpath sub commands available in virt-install since 3.0.0.
Recent (22.04+) Ubuntu host required.

Alternatively, instead of adding an xpath sub command to --network,
the same result can be achieved by adding to `virt-install` the argument:

    --xml ./devices/interface/@trustGuestRxFilters=yes

but this incantation is longer since the full path has to be specified,
while also adding the possibility to mismatch the affected interface, if
more than one were to be created.